+++ spec/lib/karafka/web/ui/pro/controllers/explorer_spec.rb
@@ -108,6 +108,25 @@ RSpec.describe_current do
       end
     end
 
+    context 'when we view first page from a topic with one partition with transactional data' do
+      before do
+        produce_many(topic, Array.new(30, '1'), type: :transactional)
+        get "explorer/#{topic}"
+      end
+
+      it do
+        expect(response).to be_ok
+        expect(body).to include(breadcrumbs)
+        expect(body).to include(pagination)
+        expect(body).to include("#{topic}/0/6")
+        expect(body).to include("#{topic}/0/29")
+        expect(body).to include(compacted_or_transactional_offset)
+        expect(body).not_to include("#{topic}/0/30")
+        expect(body).not_to include("#{topic}/0/4")
+        expect(body).not_to include(support_message)
+      end
+    end
+
     context 'when we view last page from a topic with one partition with data' do
       before do
         produce_many(topic, Array.new(30, '1'))
@@ -234,6 +253,24 @@ RSpec.describe_current do
         expect(body).to include(breadcrumbs)
         expect(body).to include('Watermark offsets')
         expect(body).to include('high: 1')
+        expect(body).to include('low: 0')
+        expect(body).not_to include(no_data)
+        expect(body).not_to include(pagination)
+        expect(body).not_to include(support_message)
+      end
+    end
+
+    context 'when only single transactional result in a given partition is present' do
+      before do
+        produce(topic, '1', type: :transactional)
+        get "explorer/#{topic}/0"
+      end
+
+      it do
+        expect(response).to be_ok
+        expect(body).to include(breadcrumbs)
+        expect(body).to include('Watermark offsets')
+        expect(body).to include('high: 2')
         expect(body).to include('low: 0')
         expect(body).not_to include(no_data)
         expect(body).not_to include(pagination)