+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - [Enhancement] Support transactional producer usage with Web UI.
 - [Fix] Fix a bug where critical errors (like `IRB::Abort`) would not abort the ongoing transaction.
 - [Fix] Prevent a scenario where an ongoing transactional producer would have stats emitted and an error that could not have been dispatched because of the transaction, creating a dead-lock.
+- [Fix] Make sure that the `recent` displays the most recent non-compacted, non-system message.
+- [Fix] Improve the `recent` message display to compensate for aborted transactions.
+- [Fix] Fix `ReferenceError: response is not defined` that occurs when Web UI returns refresh non 200.
 
 ## 0.7.7 (2023-10-20)
 - [Fix] Remove `thor` as a CLI engine due to breaking changes.